lab3-4/homework/A2/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "protobuf>=5.29.4",
]
